requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
Jinja2>=3.1.4